requirements.txt
appdirs==1.4.3
boto==2.46.1
Django==1.11
django-ckeditor==5.2.1
django-embed-video==1.1.0
django-environ==0.4.2
django-solo==1.1.2
django-storages-redux==1.3.2
greenlet==0.4.12
gunicorn==19.7.1
msgpack-python==0.4.8
neovim==0.1.13
olefile==0.44
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
